put memory functions in a seperate script
